add QDataStream << >> of map

+++ Server/12308_Server/include/vector.hpp
@@ -498,7 +498,7 @@ template<class T>
 QDataStream& operator << (QDataStream& out, const ttd::vector<T>&v)
 {
 	out << (quint16)v.size();
-	for (int i = 0; i < v.size(); ++i)
+	for (size_t i = 0; i < v.size(); ++i)
 		out << v[i] ;
 	return out;
 }
